# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "RGBSI Demo App",
    "description": "description of RGBSI Demo App",
    "version": "0.0.1",
    "contact": {
      "email": "deepakmp444@gmail.com",
      "url": "https://deepakkumar.work"
    }
  },
  "servers": [
    {
      "url": "{protocol}://localhost:5001",
      "description": "for local host",
      "variables": {
        "version": {
          "enum": ["v1", "v2"],
          "default": "v1"
        },
        "protocol": {
          "enum": ["http", "https"],
          "default": "http"
        }
      }
    }
  ],
  "components": {
...
    "securitySchemes": {
      "BasicAuth": {
        "type": "http",
        "scheme": "basic"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "JSESSIONID"
      }
    }
  },
  "paths": {
...
    "/admin/api/v1": {
      "get": {
        "tags": ["Query String"],
        "summary": "Get users",
        "parameters": [
          {
            "in": "query",
            "name": "q1",
            "schema": {
              "type": "string"
            },
            "description": "Query parameter q1"
          },
          {
            "in": "query",
            "name": "q2",
            "schema": {
              "type": "string"
            },
            "description": "Query parameter q2"
          }
        ],
        "responses": {
          "200": {
            "description": "Success response"
          }
        }
      }
    }
  }
}